backend/requirements.txt
fastapi
uvicorn
motor
python-decouple
pydantic
bson
google-genai
passlib[bcrypt]
mangum